requirements.txt
fastapi
uvicorn
pydantic
pytest
python-dotenv 
sqlalchemy
python_jose
requests